securesite/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-axes
mysqlclient
django-requestlogs-modified-0.3.1.tar.gz
